01.03 21:12 센서 허브 완료

+++ nursery/sens_hub/Core/Inc/sens.h
@@ -6,6 +6,6 @@
 extern uint8_t tds_flag; 		 // adc1 업데이트 플래그
 extern uint8_t ph_flag;  		// adc2 업데이트 플래그
 extern uint32_t tds_value; 		//adc1 버퍼
-extern uint32_t ph_value; 		//adc2 버퍼
+extern float ph_value; 		//adc2 버퍼
 
 #endif
